Start of big refactor. Untested code.
