+++ lib/board/pyboard.js
@@ -45,6 +45,7 @@ export default class Pyboard {
     this.logger = new Logger('Pyboard')
     this.config = Config.constants()
     this.refreshConfig()
+    this.configured_address = null
     this.address = null
   }
 
@@ -59,17 +60,17 @@ export default class Pyboard {
         timeout: _this.settings.timeout,
         ctrl_c_on_connect: _this.settings.ctrl_c_on_connect
       }
-      if(!_this.settings.auto_connect){
-        _this.address = _this.settings.address
-      }
+      // if(!_this.settings.auto_connect){
+      _this.configured_address = _this.settings.address
+      // }
       if(cb) cb()
     })
 
   }
-
-  setAddress(address){
-    this.address = address
-  }
+  // 
+  // setAddress(address){
+  //   this.address = address
+  // }
 
   getCallbacks(){
     return [this.onmessage,this.onerror,this.ontimeout,this.onmessage]
