+++ lib/main.js
@@ -1,8 +1,8 @@
 'use babel';
 
 import Pyboard from './board/pyboard';
-import Pymakr from './pymakr';
-import PanelView from './main/panel-view';
+import Pymakr from './main/pymakr';
+import PanelView from './views/panel-view';
 import { CompositeDisposable } from 'atom';
 import Config from './config.js'
 import SettingsWrapper from './main/settings-wrapper';
@@ -12,12 +12,22 @@ export default {
 
   activate(state) {
     var _this = this
+    console.log("Activating plugin")
     this.settings = new SettingsWrapper(function(settings){
       _this.pyboard = new Pyboard(settings)
+      console.log("Creating view")
       _this.view = new PanelView(_this.pyboard,settings,state.viewState)
+
+      console.log("Adding panel...")
+      _this.view.addPanel()
+
+      console.log("Building view")
+      _this.view.build()
+
+      console.log("Creating pymakr")
       _this.pymakr = new Pymakr(state.viewState,_this.pyboard,_this.view,settings)
 
-      _this.pymakr.addPanel()
+
 
       // Events subscribed to in atom's system can be easily cleaned up with a CompositeDisposable
       _this.subscriptions = new CompositeDisposable();
