+++ lib/main/extras.js
@@ -1,0 +1,95 @@
+'use babel';
+import Logger from '../helpers/logger.js'
+import ApiWrapper from '../main/api-wrapper.js';
+var EventEmitter = require('events');
+
+export default class Extras extends EventEmitter  {
+
+  constructor(pymakr,pyboard,terminal,settings){
+    this.pymakr = pymakr
+    this.pyboard = pyboard
+    this.terminal = terminal
+    this.settings = settings
+    this.api = new ApiWrapper(settings)
+    this.logger = new Logger('Extras')
+  }
+
+  getWifiMac(){
+    var _this = this
+    if(!this.pyboard.connected){
+      this.terminal.writeln("Please connect to your device")
+      return
+    }
+
+    var command = "from network import WLAN; from binascii import hexlify; from os import uname; wlan = WLAN(); mac = hexlify(wlan.mac()).decode('ascii'); device = uname().sysname;print('WiFi AP SSID: %(device)s-wlan-%(mac)s' % {'device': device, 'mac': mac[len(mac)-4:len(mac)]})"
+    _this.pyboard.send_wait_for_blocking(command+'\n\r',command,function(err){
+      if(err){
+        _this.logger.error("Failed to send command: "+command)
+      }
+    },1000)
+  }
+
+  getSerial(){
+    var _this = this
+    this.terminal.enter()
+
+    PySerial.list(this.settings,function(list,manufacturers){
+      _this.terminal.writeln("Found "+list.length+" serialport"+(list.length == 1 ? "" : "s"))
+      for(var i=0;i<list.length;i++){
+        var name = list[i]
+        var text = name + " (" + manufacturers[i]+ ")"
+        if(i==0){
+          _this.api.writeToCipboard(name)
+          text += " (copied to clipboard)"
+        }
+
+        _this.terminal.writeln(text)
+      }
+    })
+  }
+
+  getVersion(){
+    var _this = this
+    if(!this.pyboard.connected){
+      this.terminal.writeln("Please connect to your device")
+      return
+    }
+    var command = "import os; os.uname().release\r\n"
+    this.pyboard.send_wait_for_blocking(command,command,function(err){
+      if(err){
+        _this.logger.error("Failed to send command: "+command)
+      }
+    })
+  }
+
+  writeHelpText(){
+    var lines = []
+
+    this.terminal.enter()
+    this.terminal.write(this.config.help_text)
+
+    if(this.pyboard.connected){
+      this.logger.verbose("Write prompt")
+      this.terminal.writePrompt()
+    }
+  }
+
+  // VSCode only
+  writeGetStartedText(){
+    var _this = this
+    this.terminal.enter()
+    this.terminal.write(this.config.start_text)
+
+    Pyserial.list(this.settings,function(list){
+      if(list.length > 0){
+        _this.terminal.writeln("Here are the devices you've connected to the serial port at the moment:")
+        _this.getSerial()
+      }else if(this.pyboard.connected){
+        this.terminal.writeln()
+        this.terminal.writePrompt()
+      }
+    })
+  }
+
+
+}
