+++ lib/main/snippets.js
@@ -2,7 +2,7 @@
 
 import ApiWrapper from '../main/api-wrapper.js';
 import Utils from '../helpers/utils.js';
-import ConfigSnippets from '../config-snippets.js';
+import ConfigSnippets from '../../snippets/index.js';
 var fs = require('fs');
 var slugify = require('slugify')
 
@@ -14,9 +14,9 @@ export default class Snippets {
     this.arraylist = ConfigSnippets.defaults().files
     this.hashlist = {}
     this.custom_snippets_index = {}
-    this.package_path = this.api.getPackageSrcPath()
+    this.package_path = this.api.getPackagePath()
     this.project_path = this.api.getProjectPath()
-    this.snippets_path = this.package_path + "snippets/"
+    this.snippets_path = this.package_path + "/snippets/"
     this.custom_snippets_path = this.api.getIDEPath() + "/pymakr_snippets/"
     this.custom_snippets_index_path = this.custom_snippets_path + "index.json"
