+++ lib/main/terminal.js
@@ -10,7 +10,7 @@ export default class Term {
 
   constructor(cb,element,wrapper_element,pyboard,settings) {
     this.shellprompt = '>>> ';
-    this.element = element
+    this.element = element // get original dom element from jquery element
     this.wrapper_element = wrapper_element
     this.pyboard = pyboard
     this.element = element
@@ -36,13 +36,15 @@ export default class Term {
     })
 
     // for copy-paste with cmd key
-    this.element.addEventListener("keydown",function(e) {
+    this.element.on("keydown",function(e) {
       if (_this.isActionKey(e)) {
         _this.termKeyPress("",e)
       }
     })
 
-    this.xterm.open(element,true);
+    console.log(this.element)
+
+    this.xterm.open(this.element,true);
   }
 
   initResize(resizer){
@@ -54,9 +56,11 @@ export default class Term {
     var startTermHeight = 0
     var lineHeight = 0
     var currentRows = startRows
+    console.log(document)
 
     function onMouseDown(e){
       startY = e.clientY
+      console.log(document)
       startHeight = parseInt(document.defaultView.getComputedStyle(_this.wrapper_element).height, 10)
       startTermHeight = parseInt(document.defaultView.getComputedStyle(_this.element).height, 10)
       if(lineHeight == 0){
@@ -89,7 +93,7 @@ export default class Term {
       document.documentElement.removeEventListener('mouseup',stopDrag,false)
     }
 
-    resizer.addEventListener('mousedown',onMouseDown,false)
+    resizer.on('mousedown',onMouseDown,false)
   }
 
   setHeight(rows){
