+++ lib/views/overlay-view.html
@@ -1,0 +1,13 @@
+<div id="pymakr-overlay-wrapper">
+  <div id="wrapper-close"></div>
+  <div id="pymakr-overlay-nav">
+    <div class="snippets"></div>
+    <div class="config"></div>
+    <div class="pybytes"></div>
+  </div>
+  <div id="pymakr-overlay-content-wrapper">
+    <div id="wrapper-config" class="hidden"></div>
+    <div id="wrapper-pybytes" class="hidden"></div>
+    <div id="wrapper-snippets"></div>
+  </div>
+</div>
