+++ lib/views/panel-view.html
@@ -1,0 +1,26 @@
+
+<div id="pymakr" class="pymakr pymakr-open">
+  <div id="pymakr-resizer"></div>
+  <div id="pymakr-overlay-contents"></div>
+
+  <div id="pycom-top-bar">
+      <div id="pymakr-title">Starting...</div>
+      <div id="pymakr-buttons">
+        <button id="more">
+          <span class="fa down fa-chevron-down"></span><span class="fa up fa-chevron-up"></span> More
+          <div class="pymakr-subnav"></div>
+        </button>
+        <button id="close"><span class="fa down fa-chevron-down"></span></button>
+        <button id="settings">
+          <span class="fa fa-cog"></span> Settings
+          <div class="pymakr-subnav"></div>
+        </button>
+      </div>
+  </div>
+
+  <div id="pymakr-overlay"></div>
+
+  <div id="pymakr-terminal"></div>
+
+
+</div>
