+++ lib/views/panel-view.js
@@ -2,7 +2,7 @@
 
 import { Pane } from 'atom';
 import '../../node_modules/xterm/dist/addons/fit/fit.js';
-import Term from './terminal';
+import Term from '../main/terminal';
 import Pyserial from '../connections/pyserial';
 import ApiWrapper from '../main/api-wrapper.js';
 import Logger from '../helpers/logger.js'
@@ -26,55 +26,42 @@ export default class PanelView extends EventEmitter {
     this.settings = settings
     this.visible = true
     this.api = new ApiWrapper()
+    this.package_folder = this.api.getPackageSrcPath()
     this.logger = new Logger('PanelView')
     this.overlay = new OverlayView(this,settings)
     this.feedback_popup_seen = serializedState && 'feedbackPopupSeen' in serializedState && serializedState.feedbackPopupSeen
 
+    var html = fs.readFileSync(_this.package_folder + '/views/panel-view.html')
+    this.main_el = document.createElement('div');
+    this.main_el.insertAdjacentHTML('beforeend',html.toString())
+    console.log(html)
+    console.log(this.main_el)
+  }
+
+  build(){
     // main element
-    this.element = document.createElement('div');
-    this.element.classList.add('pymakr');
-    this.element.classList.add('pymakr-open');
-
-    // resize grip
-    this.resizer = document.createElement('div');
-    this.resizer.classList.add('pymakr-resizer');
-    this.element.appendChild(this.resizer)
-
-    // overlay contents
-    this.overlay_contents = document.createElement('div');
-    this.overlay_contents.classList.add('pymakr-overlay-contents');
-    this.element.appendChild(this.overlay_contents)
+    this.element = $('#pymakr')
+    this.resizer = $('#pymakr-resizer')
+    this.overlay_contents = $('#pymakr-overlay-contents')
+    this.topbar = $('#pycom-top-bar')
+    this.title = $('#pymakr-title')
+    this.buttons = $('#pymakr-buttons')
+    this.title = $('#pymakr-title')
+    this.button_more_tab = $('#pymakr-more-subnav')
+    this.overlay_wrapper = $('#pymakr-overlay')
+    this.terminal_element = $('#pymakr-terminal')
+    this.button_more = $('#pymakr-buttons #more')
+    this.button_more_sub = $('#pymakr-buttons #more .pymakr-subnav')
+    this.button_close = $('#pymakr-buttons #close')
+    this.button_settings = $('#pymakr-buttons #settings')
+    this.button_settings_sub = $('#pymakr-buttons #settings .pymakr-subnav')
 
     this.showFeedbackPopup()
 
-    // top bar
-    var topbar = document.createElement('div');
-    topbar.classList.add('pycom-top-bar');
-    this.topbar = topbar
-    this.title = topbar.appendChild(document.createElement('div'));
-    this.title.classList.add('pymakr-title');
-    this.title.innerHTML = 'Starting...';
-
-    // button wrapper
-    var buttons_wrapper = topbar.appendChild(document.createElement('div'));
-    buttons_wrapper.classList.add('pymakr-buttons')
-    this.buttons = buttons_wrapper
-
     // build all buttons inside wrapper
-    this.buildButtons(buttons_wrapper)
+    this.buildButtons(this.buttons)
 
-    this.button_more_tab = this.button_more.appendChild(document.createElement('div'))
-    this.button_more_sub.classList.add('pymakr-subnav');
-
-    // overlay
-    this.overlay_wrapper = document.createElement('div');
-    this.overlay_wrapper.classList.add('pymakr-overlay');
-    this.element.appendChild(this.overlay_wrapper)
-
-    // append topbar to main element
-    this.element.appendChild(topbar);
-
-    this.overlay.build()
+    this.overlay.build(this.overlay_contents)
 
     // creates the terminal elements and bindings
     this.buildTerminal()
@@ -87,7 +74,6 @@ export default class PanelView extends EventEmitter {
     this.settings.registerProjectChangeWatcher(function(path){
       _this.setProjectName(path)
     })
-
   }
 
   openSnippet(s){
@@ -97,7 +83,7 @@ export default class PanelView extends EventEmitter {
   showFeedbackPopup(){
     var _this = this
     if(!this.feedback_popup_seen){
-      this.feedback_question = this.element.appendChild(document.createElement('div'))
+      this.feedback_question = document.createElement('div')
       this.feedback_question.classList.add('pymakr-feedback');
       this.feedback_question.innerHTML = "<h2>Hi Pymakr User!</h2> "
       this.feedback_question.innerHTML += "We are working on ideas for Pymakr 2.0 and would love your feedback! "
@@ -113,6 +99,8 @@ export default class PanelView extends EventEmitter {
       this.feedback_question_close = this.feedback_question.appendChild(document.createElement('div'))
       this.feedback_question_close.classList.add('close-button');
       this.feedback_question_close.innerHTML = "x"
+
+      this.element.append(this.feedback_question)
 
       this.feedback_question_close.onclick = function(){
         _this.feedback_question.classList.add("hidden")
@@ -133,22 +121,7 @@ export default class PanelView extends EventEmitter {
 
   }
 
-  buildButtons(buttons_wrapper){
-
-
-    this.button_more = document.createElement('button');
-    this.button_more.innerHTML = '<span class="fa down fa-chevron-down"></span><span class="fa up fa-chevron-up"></span> More';
-
-    this.button_more_sub = this.button_more.appendChild(document.createElement('div'))
-    this.button_more_sub.classList.add('pymakr-subnav');
-
-    this.button_close = buttons_wrapper.appendChild(document.createElement('button'));
-    this.button_close.innerHTML = '<span class="fa fa-chevron-down"></span>';
-
-    this.button_settings = buttons_wrapper.appendChild(document.createElement('button'));
-    this.button_settings.innerHTML = '<span class="fa fa-cog"></span> Settings';
-    this.button_settings_sub = this.button_settings.appendChild(document.createElement('div'))
-    this.button_settings_sub.classList.add('pymakr-subnav');
+  buildButtons(){
 
     this.option_global_settings = this.createButton('global_settings','div','','Global settings','',this.button_settings_sub)
     this.option_project_settings = this.createButton('project_settings','div','','Project settings','',this.button_settings_sub)
@@ -167,26 +140,22 @@ export default class PanelView extends EventEmitter {
     this.option_get_help = this.createButton('','div','','Help','')
     this.option_get_snippets = this.createButton('','div','','Snippets','') //inactive for 1.5 version (see snippets branch for latest working version)
 
-    buttons_wrapper.appendChild(this.button_more)
   }
 
   buildTerminal(){
     var _this = this
-    // terminal UI elements
-    this.terminal_el = document.createElement('div');
-    this.terminal_el.id = "pymakr-terminal"
-    this.element.appendChild(this.terminal_el);
 
     // terminal resize functionality
     var erd = ElementResize();
-    erd.listenTo(this.terminal_el,function(element){
+    erd.listenTo(this.terminal_element,function(element){
       if(_this.visible){
           _this.setPanelHeight()
       }
     })
 
     // create terminal
-    this.terminal = new Term(null,this.terminal_el,_this.element,this.pyboard,null)
+    console.log(this.terminal_element)
+    this.terminal = new Term(null,_this.terminal_element,_this.element,this.pyboard,null)
     this.terminal.initResize(_this.resizer)
     this.terminal.setOnMessageListener(function(input){
       _this.emit('user_input',input)
@@ -294,10 +263,9 @@ export default class PanelView extends EventEmitter {
 
 
     // 'click to connect' feature on complete terminal element
-    this.terminal_el.onclick = function(){
+    this.terminal_element.onclick = function(){
       _this.emit('terminal_click')
     }
-
   }
 
   createButton(id,type,icon,text,classname,parent){
@@ -310,7 +278,8 @@ export default class PanelView extends EventEmitter {
       }
     }
 
-    button = parent.appendChild(document.createElement(type));
+    button = document.createElement(type)
+    parent.append(button);
     button.innerHTML = '<span id="pymakr-'+id+'" class="fa fa-'+icon+'"></span> '+text;
     if(classname && classname != ''){
       button.classList.add('pymakr-'+classname);
@@ -397,6 +366,7 @@ export default class PanelView extends EventEmitter {
 
   // UI Stuff
   addPanel(){
+    console.log("Adding bottom panel to html")
     this.api.addBottomPanel(
       {
         item: this.getElement(),
@@ -408,7 +378,7 @@ export default class PanelView extends EventEmitter {
 
   setPanelHeight(height){
     if(!height){
-      height = (this.terminal_el.offsetHeight + 25)
+      height = (this.terminal_element.offsetHeight + 25)
     }
     this.element.style.height = height + "px"
   }
@@ -452,7 +422,7 @@ export default class PanelView extends EventEmitter {
   }
 
   getElement() {
-    return this.element;
+    return this.main_el;
   }
 
 }
