+++ lib/views/snippets-view.html
