+++ lib/views/snippets-view.js
@@ -2,7 +2,7 @@
 
 import { Pane } from 'atom';
 import '../../node_modules/xterm/dist/addons/fit/fit.js';
-import Term from './terminal';
+import Term from '../main/terminal';
 import Pyserial from '../connections/pyserial';
 import ApiWrapper from '../main/api-wrapper.js';
 import Logger from '../helpers/logger.js'
@@ -24,14 +24,13 @@ export default class SnippetsView extends EventEmitter {
     this.package_folder
   }
 
-  build(){
+  build(root_element){
     var _this = this
     $(document).ready(function(){
 
-      var snippets_html = fs.readFileSync(_this.package_folder + 'main/snippets-view.html')
-      $('.pymakr-overlay-content-wrapper').append(snippets_html.toString())
+      var snippets_html = fs.readFileSync(_this.package_folder + '/views/snippets-view.html')
+      root_element.append(snippets_html.toString())
 
-      _this.content = $('#snippets')
       _this.list = $('#snippets-list')
       _this.snippet_name = $('#snippet-name')
       _this.content_right = $('#snippets-right')
