+++ snippets/lte-get-imei.py
