+++ snippets/lte-nb-iot.py
