+++ snippets/wifi-simple.py
