+++ snippets/wifi-static-ip.py
